black

+++ fiduswriter/ojs/views.py
@@ -116,10 +116,7 @@ def open_revision_doc(request, submission_id, version):
         return HttpResponse("No access", status=403)
     if (
         rev.document.owner != user
-        and AccessRight.objects.filter(
-            document=rev.document, user=user
-        ).count()
-        == 0
+        and AccessRight.objects.filter(document=rev.document, user=user).count() == 0
     ):
         # The user to be logged in is neither the editor (owner of doc), nor
         # has he access rights to the doc. We prohibit access.
@@ -152,9 +149,7 @@ def check_revision_doc(request, submission_id, version):
         except ValueError:
             is_editor = 0
 
-        user = find_user(
-            journal_id, submission_id, version, user_id, is_editor
-        )
+        user = find_user(journal_id, submission_id, version, user_id, is_editor)
 
         # Validate if user exists
         if not user:
@@ -187,9 +182,7 @@ def get_doc_info(request):
         document = Document.objects.get(id=document_id)
         if (
             document.owner != request.user
-            and AccessRight.objects.filter(
-                document=document, user=request.user
-            ).count()
+            and AccessRight.objects.filter(document=document, user=request.user).count()
             == 0
         ):
             # Access forbidden
@@ -205,12 +198,7 @@ def get_doc_info(request):
             reviewer = revision.reviewer_set.filter(user=request.user).first()
             if reviewer:
                 user_role = "reviewer"
-            elif (
-                revision.submission.author_set.filter(
-                    user=request.user
-                ).count()
-                > 0
-            ):
+            elif revision.submission.author_set.filter(user=request.user).count() > 0:
                 # User with author role but not submission submitter as sub-author
                 user_role = (
                     "author"
@@ -227,9 +215,7 @@ def get_doc_info(request):
             response["submission"]["status"] = "submitted"
             response["submission"]["submission_id"] = revision.submission.id
             response["submission"]["version"] = revision.version
-            response["submission"][
-                "journal_id"
-            ] = revision.submission.journal_id
+            response["submission"]["journal_id"] = revision.submission.journal_id
             response["submission"]["user_role"] = user_role
             if reviewer and reviewer.method == "doubleanonymous":
                 response["submission"]["contributors"] = {}
@@ -446,9 +432,7 @@ def remove_reviewer(request, submission_id, version):
         response["error"] = "Unknown reviewer"
         status = 403
         return JsonResponse(response, status=status)
-    AccessRight.objects.filter(
-        document=revision.document, user=reviewer.user
-    ).delete()
+    AccessRight.objects.filter(document=revision.document, user=reviewer.user).delete()
     reviewer.delete()
     return JsonResponse(response, status=status)
 
@@ -492,9 +476,7 @@ def create_copy(request, submission_id):
             for editor in editors:
                 if str(editor.ojs_jid) in granted_user_ids:
                     role = int(editor.role)
-                    rights = constants.EDITOR_ROLE_STAGE_RIGHTS[role][
-                        new_version_stage
-                    ]
+                    rights = constants.EDITOR_ROLE_STAGE_RIGHTS[role][new_version_stage]
                     AccessRight.objects.create(
                         document=revision.document,
                         holder_obj=editor.user,
@@ -572,9 +554,7 @@ def add_editor(request, submission_id):
                     ).first()
                     if access_right is None:
                         role = int(editor.role)
-                        rights = constants.EDITOR_ROLE_STAGE_RIGHTS[role][
-                            int(stage_id)
-                        ]
+                        rights = constants.EDITOR_ROLE_STAGE_RIGHTS[role][int(stage_id)]
                         access_right = AccessRight(
                             document=revision.document,
                             holder_obj=editor.user,
@@ -613,9 +593,7 @@ def remove_editor(request, submission_id):
         status = 403
         return JsonResponse(response, status=status)
 
-    revisions = models.SubmissionRevision.objects.filter(
-        submission_id=submission_id
-    )
+    revisions = models.SubmissionRevision.objects.filter(submission_id=submission_id)
     if revisions is not None:
         for revision in revisions:
             AccessRight.objects.filter(
@@ -661,9 +639,7 @@ def add_author(request, submission_id):
 
     # create access_rights for existing revisions
     # get ids of stages access granted
-    revisions = models.SubmissionRevision.objects.filter(
-        submission_id=submission_id
-    )
+    revisions = models.SubmissionRevision.objects.filter(submission_id=submission_id)
     if revisions is not None:
         for revision in revisions:
             version = revision.version.split(".")
@@ -723,9 +699,7 @@ def remove_author(request, submission_id):
         status = 403
         return JsonResponse(response, status=status)
 
-    revisions = models.SubmissionRevision.objects.filter(
-        submission_id=submission_id
-    )
+    revisions = models.SubmissionRevision.objects.filter(submission_id=submission_id)
     if revisions is not None:
         for revision in revisions:
             AccessRight.objects.filter(
